backend/requirements.txt
flask
flask-cors
pytesseract
Pillow
PyMuPDF
gunicorn